+++ crates/runtime/src/flight/flightsql/statement_query.rs
@@ -1,6 +1,7 @@
 use std::sync::Arc;
 
 use arrow_flight::{
+    flight_service_server::FlightService,
     sql::{self, ProstMessageExt},
     FlightDescriptor, FlightEndpoint, FlightInfo, Ticket,
 };
@@ -9,7 +10,7 @@ use tonic::{Request, Response, Status};
 
 use crate::flight::{to_tonic_err, Service};
 
-/// Get a FlightInfo for executing a SQL query.
+/// Get a `FlightInfo` for executing a SQL query.
 pub(crate) async fn get_flight_info(
     flight_svc: &Service,
     query: sql::CommandStatementQuery,
@@ -38,4 +39,21 @@ pub(crate) async fn get_flight_info(
         .with_total_records(num_rows.try_into().map_err(to_tonic_err)?);
 
     Ok(Response::new(info))
+}
+
+pub(crate) async fn do_get(
+    flight_svc: &Service,
+    ticket: sql::TicketStatementQuery,
+) -> Result<Response<<Service as FlightService>::DoGetStream>, Status> {
+    let datafusion = Arc::clone(&flight_svc.datafusion);
+    tracing::trace!("do_get_statement: {ticket:?}");
+    match std::str::from_utf8(&ticket.statement_handle) {
+        Ok(sql) => {
+            let output = Service::sql_to_flight_stream(datafusion, sql.to_owned()).await?;
+            Ok(Response::new(
+                Box::pin(output) as <Service as FlightService>::DoGetStream
+            ))
+        }
+        Err(e) => Err(Status::invalid_argument(format!("Invalid ticket: {e:?}"))),
+    }
 }
